% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{contract_idINS}
\alias{contract_idINS}
\title{convertit les idINS longs en courts}
\usage{
contract_idINS(idINS)
}
\arguments{
\item{idINS}{les id longs}
}
\value{
des id courts (100000*x/r + y/r)
}
\description{
convertit les idINS longs en courts
}
