% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{coord2idINS}
\alias{coord2idINS}
\title{Creates identifiers for (x,y) coordinates.
Used to work with crs 3035}
\usage{
coord2idINS(x, y = NULL, resolution = 200)
}
\arguments{
\item{x}{matrix of coordinates (first column = x, second = y). Or a vector of x coordinates.}

\item{y}{default to NULL (if x is a matrix). Or a vector of y coordinates.}

\item{resolution}{default to 200. Size of raster cells which are identified.}
}
\description{
Creates identifiers for (x,y) coordinates.
Used to work with crs 3035
}
