% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dt.r
\name{dt2r}
\alias{dt2r}
\title{Crée un raster à partir d'un data.table avec idINS.}
\usage{
dt2r(dt, resolution = NULL, idINS = "idINS")
}
\arguments{
\item{dt}{data.table avec idINS.}

\item{resolution}{résolution du raster.}

\item{idINS}{nom de la variable idINS, par défaut "idINS".}
}
\description{
Crée un raster à partir d'un data.table avec idINS.
}
