% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dt.r
\name{dt2stars}
\alias{dt2stars}
\title{Transform a data.table with idINS to a star object.
Assumes crs 3035}
\usage{
dt2stars(data, idINS = "idINS")
}
\arguments{
\item{data}{a data.table with an idINS column}

\item{idINS}{name of the idINS column. Default to idINS.}
}
\description{
Transform a data.table with idINS to a star object.
Assumes crs 3035
}
