% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{expand_idINS}
\alias{expand_idINS}
\title{convertit les idINS courts en longs}
\usage{
expand_idINS(ids, resolution = 200)
}
\arguments{
\item{ids}{les id courts}

\item{resolution}{la résolution (defaut 200m)}
}
\value{
un vecteur d'idINS longs (ex. r200N2051600E4258800)
}
\description{
convertit les idINS courts en longs
}
