% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{f2si2}
\alias{f2si2}
\title{Mise en forme des nombres (vient de f.communs)}
\usage{
f2si2(number, rounding = TRUE, digits = 1, unit = "median")
}
\arguments{
\item{number}{nombre à mettre en forme}

\item{rounding}{logique. TRUE par défaut.}

\item{digits}{nombre de chiffre après la virgule.}

\item{unit}{méthode de choix de l'unité ("median" par défaut)}
}
\description{
Mise en forme des nombres (vient de f.communs)
}
