% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_Raster.r
\name{geom_Raster}
\alias{geom_Raster}
\title{ajoute une couche raster à un ggplot}
\usage{
geom_Raster(raster, mapping = NULL, ..., long = FALSE, style = "cont", k = 5)
}
\arguments{
\item{raster}{le raster (package raster) au format rasterlayer, rasterbrick ou rasterstack à afficher}

\item{mapping}{un mapping supplémentaire (définit la colonne employé pour la couleur)}

\item{...}{autres paramètres passés à geom_tile}

\item{long}{permet de passer des raster au format long si on veut pour les différentes couches du raster}

\item{style}{kmeans}

\item{k}{nombre de groupes (kmeans)}
}
\value{
une couche ggplot2
}
\description{
geom_raster demande de préciser le système de coordonnées (aes(x=lon, y=lat)). \code{geom_Ratser} (avec un R majuscule)
permet de s'en passer. Attention le système de coordonnées est uniquement 3035.
}
