% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getresINS}
\alias{getresINS}
\title{Récupère la résolution dans idINS.}
\usage{
getresINS(dt, idINS = "idINS")
}
\arguments{
\item{dt}{data.table ou df.}

\item{idINS}{variable contenant idINS, par défaut, idINS.}
}
\description{
Récupère la résolution dans idINS.
}
