% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{idINS2coord}
\alias{idINS2coord}
\title{Retrieves (x, y) coordinates from idINS.
Used to work with crs 3035}
\usage{
idINS2coord(ids, stop_if_res_not_cst = TRUE, resolution = 200)
}
\arguments{
\item{ids}{character vector of idINS}

\item{stop_if_res_not_cst}{default to TRUE. Refuse to convert id with different resolutions}
}
\description{
Retrieves (x, y) coordinates from idINS.
Used to work with crs 3035
}
