% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{idINS2dist}
\alias{idINS2dist}
\title{Calculate euclidean distance between to long or short idINS, in meter}
\usage{
idINS2dist(fromidINS, toidINS, resolution = NULL)
}
\arguments{
\item{fromidINS}{character vector of starting idINS}

\item{toidINS}{character vector of ending idINS}

\item{resolution}{default to NULL. Set if no resolution is provided in idINS}
}
\description{
Calculate euclidean distance between to long or short idINS, in meter
}
