% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{idINS2point}
\alias{idINS2point}
\title{Récupère les coordonnées X et Y de idINS court ou long.}
\usage{
idINS2point(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}

\item{idINS}{le champ qui contient l'idINS court (défaut "idINS")}
}
\description{
Récupère les coordonnées X et Y de idINS court ou long.
}
