% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r, R/short idINS.R
\name{lidINS2dist}
\alias{lidINS2dist}
\title{Calculate euclidean distance between to long idINS, in meter}
\usage{
lidINS2dist(fromidINS, toidINS, resolution = NULL)

lidINS2dist(fromidINS, toidINS, resolution = NULL)
}
\arguments{
\item{fromidINS}{character vector of starting idINS}

\item{toidINS}{character vector of ending idINS}

\item{resolution}{default to NULL. Set if no resolution is provided in idINS}
}
\description{
Calculate euclidean distance between to long idINS, in meter

Calculate euclidean distance between to long or short idINS, in meter
}
