% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{lidINS2lonlat}
\alias{lidINS2lonlat}
\title{Retrieves (lon, lat) coordinates from long idINS.}
\usage{
lidINS2lonlat(idINS, resolution = NULL)
}
\arguments{
\item{idINS}{character vector of idINS}

\item{resolution}{default to NULL. Set to convert id with different resolutions}
}
\description{
Retrieves (lon, lat) coordinates from long idINS.
}
