% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{lidINS2point}
\alias{lidINS2point}
\title{Récupère les coordonnées X et Y de idINS long}
\usage{
lidINS2point(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Récupère les coordonnées X et Y de idINS long
}
