% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{lidINS2square}
\alias{lidINS2square}
\title{Convertit l'idINS en polygone}
\usage{
lidINS2square(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Convertit l'idINS en polygone
}
