% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dt.r
\name{r2dt}
\alias{r2dt}
\title{Raster vers data.table}
\usage{
r2dt(raster, resolution = NULL, fun = mean, shortidINS = FALSE)
}
\arguments{
\item{raster}{le raster \code{RatserLayer}, \code{RatserBrick},  \code{RatserStack}}

\item{resolution}{résolution en mètre pour la transformation (200m par défaut)}

\item{fun}{une fonction pour l'agrégation (mean par défaut)}
}
\value{
un data.table avec un idINS
}
\description{
Transforme un raster::ratser en data.table
en ajoutant un idINS. Fonction inverse de dt2r
}
