% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{r2sdt}
\alias{r2sdt}
\title{Raster vers data.table avec idINS short}
\usage{
r2sdt(raster, resolution = 200, fun = mean)
}
\arguments{
\item{raster}{le raster \code{RatserLayer}, \code{RatserBrick},  \code{RatserStack}}

\item{resolution}{résolution en mètre pour la transformation (200m par défaut)}

\item{fun}{une fonction pour l'agrégation (mean par défaut)}
}
\value{
un data.table avec un idINS
}
\description{
Transforme un raster::ratser en data.table
en ajoutant un idINS. Fonction inverse de sdt2r
}
