% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{raster_ref}
\alias{raster_ref}
\title{Crée un pavage carré en tant que géométrie d'un sf à partir de idINS.}
\usage{
raster_ref(data, resolution = 200, crs = 3035)
}
\arguments{
\item{data}{sf.}

\item{resolution}{par défaut, 200.}

\item{crs}{par défaut, 3035 (European Terrestrial Reference System).}

\item{ids}{vecteur d'idINS.}
}
\description{
Crée un pavage carré en tant que géométrie d'un sf à partir de idINS.
}
