% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastermap.r
\name{rastervar}
\alias{rastervar}
\title{ratervar}
\usage{
rastervar(data, ..., fun = mean, dropth = 0, resolution = 200, idINS = "idINS")
}
\arguments{
\item{data}{données (tibble, data.frame) avec un idINS ou un sf (l'idINS sera fabriqué par projection sur une grille 3035)}

\item{...}{expressions en quasiquotation}

\item{fun}{fonction d'agrégagtion}

\item{dropth}{Elimine les extrêmes de la distribtion (un double, inférieur à 1)}

\item{resolution}{résolution par défaut 200}

\item{idINS}{nom de la variable idINS}
}
\value{

}
\description{
ratervar
}
