% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sdt2r}
\alias{sdt2r}
\title{Crée un raster à partir d'un data.table avec idINS short}
\usage{
sdt2r(dt, resolution = 200, idINS = "idINS")
}
\arguments{
\item{dt}{data.table avec idINS.}

\item{resolution}{résolution du raster.}

\item{idINS}{nom de la variable idINS, par défaut "idINS".}
}
\description{
Crée un raster à partir d'un data.table avec idINS short
}
