% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS2dist}
\alias{sidINS2dist}
\title{Calculate euclidean distance between to short idINS, in meter}
\usage{
sidINS2dist(fromidINS, toidINS, resolution = 200)
}
\arguments{
\item{fromidINS}{character vector of starting idINS}

\item{toidINS}{character vector of ending idINS}

\item{resolution}{default to NULL. Set if no resolution is provided in idINS}
}
\description{
Calculate euclidean distance between to short idINS, in meter
}
