% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS2lonlat}
\alias{sidINS2lonlat}
\title{Retrieves (lon, lat) coordinates from long idINS.}
\usage{
sidINS2lonlat(idINS, resolution = 200)
}
\arguments{
\item{idINS}{character vector of idINS}

\item{resolution}{default to NULL. Set to convert id with different resolutions}
}
\description{
Retrieves (lon, lat) coordinates from long idINS.
}
