% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS2point}
\alias{sidINS2point}
\title{Récupère les coordonnées X et Y de idINS.}
\usage{
sidINS2point(ids, resolution = 200)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Récupère les coordonnées X et Y de idINS.
}
