% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS2sf}
\alias{sidINS2sf}
\title{retourne un sf à partir d'un data.frame avec un champ idINS}
\usage{
sidINS2sf(data, resolution = 200, idINS = "idINS")
}
\arguments{
\item{data}{le data frame}

\item{resolution}{200m par défaut}

\item{idINS}{le champ qui contient l'idINS court (défaut "idINS")}
}
\value{
un sf
}
\description{
retourne un sf à partir d'un data.frame avec un champ idINS
}
