% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS2square}
\alias{sidINS2square}
\title{Convertit l'idINS en polygone}
\usage{
sidINS2square(ids, resolution = 200)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Convertit l'idINS en polygone
}
