% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{sidINS3035}
\alias{sidINS3035}
\title{Crée idINS à partir de coordonnées.}
\usage{
sidINS3035(x, y = NULL, resolution = 200)
}
\arguments{
\item{x}{Ou un vecteur de la coordonnée x, ou un df avec les colonnes x et y.}

\item{y}{vecteur de la coordonnée y, si nécessaire. NULL par défaut.}

\item{resolution}{resolution, par défaut 200.}

\item{resinstr}{faut-il inscrire la résolution dans idINS ? Par défaut TRUE.}
}
\description{
Crée idINS à partir de coordonnées.
}
