% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{stars_ref}
\alias{stars_ref}
\title{Produce a stars object with zero value on a fixed resolution}
\usage{
stars_ref(data, resolution, values = 0L)
}
\arguments{
\item{data}{spatial feature of the relevant locations}

\item{resolution}{cell size}

\item{values}{values to populate the stars values with, default to 0}
}
\description{
Produce a stars object with zero value on a fixed resolution
}
