% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dt.r, R/utils.r
\name{dt2r}
\alias{dt2r}
\title{data.table vers Raster}
\usage{
dt2r(dt, resolution = NULL, idINS = "idINS")

dt2r(dt, resolution = NULL, idINS = "idINS")
}
\arguments{
\item{dt}{data.table avec idINS.}

\item{resolution}{résolution du raster.}

\item{idINS}{nom de la variable idINS, par défaut "idINS".}
}
\value{
un raster, dans le système de coordonnées 3035
}
\description{
Transforme un data.table en raster
en ajoutant un idINS
}
