% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r, R/short idINS.R
\name{idINS2lonlat}
\alias{idINS2lonlat}
\title{Retrieves (lon, lat) coordinates from idINS.}
\usage{
idINS2lonlat(idINS, resolution = NULL)

idINS2lonlat(idINS, resolution = NULL)
}
\arguments{
\item{idINS}{character or integer vector of idINS}

\item{resolution}{default to NULL. Set to convert id with different resolutions}
}
\description{
Retrieves (lon, lat) coordinates from idINS.

Retrieves (lon, lat) coordinates from long or short idINS.
}
