% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idINS.r
\name{idINS2point}
\alias{idINS2point}
\title{Récupère les coordonnées X et Y de idINS.}
\usage{
idINS2point(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Récupère les coordonnées X et Y de idINS.
}
