% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short idINS.R
\name{idINS2square}
\alias{idINS2square}
\title{Récupère les coordonnées X et Y de idINS court ou long.}
\usage{
idINS2square(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Récupère les coordonnées X et Y de idINS court ou long.
}
