% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dt.r, R/utils.r
\name{raster_ref}
\alias{raster_ref}
\title{Title}
\usage{
raster_ref(data, resolution = 200, crs = 3035)

raster_ref(data, resolution = 200, crs = 3035)
}
\arguments{
\item{data}{sf.}

\item{resolution}{par défaut, 200.}

\item{crs}{par défaut, 3035 (European Terrestrial Reference System).}
}
\value{

}
\description{
Title

raster ref
}
